\name{testrffi-package}
\alias{testrffi-package}
\alias{testrffi}
\docType{package}
\title{Tests a package with native code}
\description{Tests a package with native code}
\details{
\tabular{ll}{
Package: \tab testrffi\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-21\cr
License: \tab GPL-2\cr
}

}
\author{
The FastR Team

Maintainer: fastr@yahoogroups.com
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{

}
