\name{testrffi}
\alias{testrffi}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
testrffi
}
\description{
Tests the R FFI interface
}
\usage{
addint(2,3)
}

\details{

}
\value{
invisble NULL
}
\references{

}
\author{
FastR Team
}
\note{

}

\seealso{
%
}
\examples{
addint(2,3)
}
