% Copyright (c) 2018, Oracle and/or its affiliates. All rights reserved.
% DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
%
% This code is free software; you can redistribute it and/or modify it
% under the terms of the GNU General Public License version 3 only, as
% published by the Free Software Foundation.
%
% This code is distributed in the hope that it will be useful, but WITHOUT
% ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
% FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
% version 3 for more details (a copy is included in the LICENSE file that
% accompanied this code).
%
% You should have received a copy of the GNU General Public License version
% 3 along with this work; if not, write to the Free Software Foundation,
% Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
%
% Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
% or visit www.oracle.com if you need additional information or have any
% questions.
\title{Tests the R FFI interface}
\name{testrffi-package}
\alias{testrffi-package}
\alias{testrffi}
\docType{package}
\description{Tests the R FFI interface via a collection of R functions that
call down to native code using .Call which then typically make R FFI upcalls to
test the individual functions. The set of tests is not comprehensive.

The individual functions are not documented here as there are far too many
and they are not intended for user-level usage.
}
\details{
\tabular{ll}{
Package: \tab testrffi\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-02-20\cr
License: \tab GPL-2\cr
}

}
\author{
The FastR Team

Maintainer: fastr@yahoogroups.com
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{
}
