\title{Tests the R FFI interface}
\name{testrffi-package}
\alias{testrffi-package}
\alias{testrffi}
\docType{package}
\description{Tests the R FFI interface via a collection of R functions that
call down to native code using .Call which then typically make R FFI upcalls to
test the individual functions. The set of tests is not comprehensive.

The individual functions are not documented here as there are far too many
and they are not intended for user-level usage.
}
\details{
\tabular{ll}{
Package: \tab testrffi\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-02-20\cr
License: \tab GPL-2\cr
}

}
\author{
The FastR Team

Maintainer: fastr@yahoogroups.com
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{
}
